Utilsを作成して処理を移動
